rewrite to preform check in popup
Previously, the check used `chrome.webRequest` to capture response
headers. This new logic preforms a separate request in the popup to
check the server headers.

Fixes #1
Fixes #3
Fixes #5
Fixes #6

The icon now defaults to blue, and isn't dynamic -- we don't know if the
site is hosted by GitHub Pages until the popup is opened.

Fixes #4